ajout du fichier Modules

+++ supportligne/database/migrations/2024_09_23_125338_create_customer_projet_table.php
@@ -6,22 +6,31 @@ use Illuminate\Database\Migrations\Migration;
 
 class CreateCustomerProjetTable extends Migration
 {
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
     public function up()
     {
         Schema::create('customer_projet', function (Blueprint $table) {
-            // Colonnes pour les relations
-            $table->unsignedBigInteger('id_customers');  // 'unsignedBigInteger' pour correspondre à la clé primaire de 'customers'
-            $table->unsignedBigInteger('id_projet');     // 'unsignedBigInteger' pour correspondre à la clé primaire de 'projet'
+            $table->unsignedInteger('id_customers');
+            $table->unsignedBigInteger('id_projet');
             
-            // Définition des clés étrangères
+            // Clés étrangères
             $table->foreign('id_customers')->references('id')->on('customers')->onDelete('cascade');
             $table->foreign('id_projet')->references('id')->on('projet')->onDelete('cascade');
             
-            // Définition de la clé primaire composite
+            // Clé primaire composite
             $table->primary(['id_customers', 'id_projet']);
         });
     }
 
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
     public function down()
     {
         Schema::dropIfExists('customer_projet');
